Crate libc_core

This crate provides the libc types.

Modules

consts

This file contains constants used in the libc crate.

elf

This module provides the libc types for ELF (Executable and Linkable Format).

epoll

This module provides the libc types for Epoll (event polling).

fcntl

This module provides the libc types for FCNTL (file control).

futex

This module provides the libc types for FUTEX (fast user-space mutex).

internal

This module provides the libc types for libc internal use.

ioctl

This module provides the libc types for IOCTL.

mman

This module provides the libc types for MMAN (memory management).

others

This module provides the libc types for Other unclassified types.

poll

This module provides the libc types for Poll (polling).

resource

This module provides the libc types for Resource (system resource management).

sched

This module provides the libc types for SCHED (scheduling).

signal

This module provides the libc types for Signal.

termios

This module provides the libc types for Termios.

time

This module provides the libc types for Time (time management).

times

This module provides the libc types for TIMES (time management).

types

This module provides the libc types for Types.

utsname

This module provides the libc types for UTSNAME (Unix Time Sharing Name).