libc_core::types

Struct StatMode

pub struct StatMode(/* private fields */);

文件的状态信息，类似于 Linux 中的 stat 结构体。

Implementations

impl StatMode

pub const NULL: Self

Null

pub const TYPE_MASK: Self

Type

pub const FIFO: Self

FIFO

pub const CHAR: Self

character device

pub const DIR: Self

directory

pub const BLOCK: Self

block device

pub const FILE: Self

ordinary regular file

pub const LINK: Self

symbolic link

pub const SOCKET: Self

socket

pub const SET_UID: Self

Set-user-ID on execution.

pub const SET_GID: Self

Set-group-ID on execution.

pub const OWNER_MASK: Self

Read, write, execute/search by owner.

pub const OWNER_READ: Self

Read permission, owner.

pub const OWNER_WRITE: Self

Write permission, owner.

pub const OWNER_EXEC: Self

Execute/search permission, owner.

pub const GROUP_MASK: Self

Read, write, execute/search by group.

pub const GROUP_READ: Self

Read permission, group.

pub const GROUP_WRITE: Self

Write permission, group.

pub const GROUP_EXEC: Self

Execute/search permission, group.

pub const OTHER_MASK: Self

Read, write, execute/search by others.

pub const OTHER_READ: Self

Read permission, others.

pub const OTHER_WRITE: Self

Write permission, others.

pub const OTHER_EXEC: Self

Execute/search permission, others.

impl StatMode

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u32

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl StatMode

pub const fn iter(&self) -> Iter<StatMode>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<StatMode>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.